Tweak: natural number check

+++ src/main/kotlin/me/xfl03/kit/util/TextUtil.kt
@@ -1,7 +1,7 @@
 package me.xfl03.kit.util
 
 fun isNaturalNumber(text: String) = try {
-    val id = text.toInt()
+    val id = text.toLong()
     id > 0
 } catch (_: Exception) {
     false
